Fix missing type cast issue

+++ src/Resources/contao/classes/CalendarEventsHelper.php
@@ -590,7 +590,7 @@ class CalendarEventsHelper
         }
 
         $eventDuration = \count(self::getEventTimestamps($objEvent));
-        $span = Calendar::calculateSpan(self::getStartDate($objEvent), self::getEndDate($objEvent)) + 1;
+        $span = (int) Calendar::calculateSpan(self::getStartDate($objEvent), self::getEndDate($objEvent)) + 1;
 
         if (1 === $eventDuration) {
             return Date::parse($dateFormat, self::getStartDate($objEvent)).($blnAppendEventDuration ? ' ('.self::getEventDuration($objEvent).')' : '');
